Fixes
Fixed some bugs in items equipment also updated us to 4.8.

Figured how to turn off readonly. now we can get rollin papers. WHO GOT THE SMOKE

+++ Scripts/Expansions/Original UO/Items/Equipment/Clothing/Surcoat.cs
@@ -1,5 +1,3 @@
-using System;
-
 namespace Server.Items
 {
     [Flipable(0x1ffd, 0x1ffe)]
@@ -27,7 +25,7 @@ namespace Server.Items
         {
             base.Serialize(writer);
 
-            writer.Write((int)0); // version
+            writer.Write(0); // version
         }
 
         public override void Deserialize(GenericReader reader)
